refactor(tauri): split 67 commands out of lib.rs into domain modules
lib.rs had grown to 8,210 lines with 75 #[tauri::command] fns mixed into
the agent-loop core. Move 67 of them into domain modules following the
existing *_commands.rs pattern (use super::*; + pub(super)):

- session_commands.rs: session/folder CRUD, paging, rewind/load/search
- project_commands.rs: project CRUD, window management, research graph
- app_commands.rs: file dialogs, notifications, onboarding, updates
- runtime_commands.rs: runtime start/stop, execution contexts, runs
- memory_commands.rs: memory file read/write/clear
- settings_commands.rs (append): auto_review / update_check toggles
- artifact_commands.rs (append): artifact list/search/read

Single-domain helpers move with their commands; cross-domain helpers
(build_project_info, list_memory_files, the GithubRelease update-check
block) stay in lib.rs. The 8 remaining commands are the agent-loop core
(send_message, enqueue_turn, stop_agent, ...) which is deeply coupled to
the chat infrastructure that lib.rs now solely owns.

Frontend invoke() names are unchanged: tauri command names derive from
the fn name only, so module paths in generate_handler! are invisible to
the webview. lib.rs drops from 8,210 to 6,163 lines.
