+++ src-tauri/src/lib.rs
@@ -18,6 +18,7 @@ use wisp_skills::SkillIndex;
 use wisp_store::{LibraryStore, Store};
 
 mod acp;
+mod app_commands;
 mod approval_commands;
 mod artifact_commands;
 mod browser_bridge;
@@ -39,10 +40,12 @@ mod library_commands;
 mod mcp_bridge;
 pub use mcp_bridge::run_mcp_bridge_cli;
 mod mcp_oauth;
+mod memory_commands;
 mod models;
 mod native_delegation;
 mod pet_commands;
 mod plugins;
+mod project_commands;
 mod project_reader;
 mod project_sync;
 mod project_transfer;
@@ -50,9 +53,11 @@ mod research_graph;
 mod resource_refs;
 mod review;
 mod run_context;
+mod runtime_commands;
 mod runtime_config_tool;
 mod runtime_launcher;
 mod seed;
+mod session_commands;
 mod session_context_tool;
 mod session_export;
 mod settings_commands;
@@ -3022,41 +3027,6 @@ fn drain_pending_notify_target(window: &tauri::Window) {
     }
 }
 
-/// Desktop notification for task status (#327). No-op while any app window is
-/// focused (the in-app UI already shows the state) or when disabled in settings.
-/// Clicking the notification navigates to the session it was about (#434) —
-/// see `pending_notify_targets`.
-#[tauri::command]
-async fn notify_user(
-    window: tauri::Window,
-    state: State<'_, AppState>,
-    title: String,
-    body: String,
-    session_id: String,
-) -> Result<(), String> {
-    if app_has_focus() || !load_notifications_enabled(&state.store).await {
-        return Ok(());
-    }
-    // Arm click-to-open before showing: on this window's next focus it jumps to
-    // the session. Skipped if the session's project can't be resolved (the
-    // notification still shows, just without navigation).
-    if let Ok(Some(project_id)) = state.store.frame_project_id(&session_id).await {
-        pending_notify_targets().lock().unwrap().insert(
-            window.label().to_string(),
-            serde_json::json!({ "projectId": project_id, "sessionId": session_id }),
-        );
-    }
-    use tauri_plugin_notification::NotificationExt;
-    window
-        .app_handle()
-        .notification()
-        .builder()
-        .title(title)
-        .body(body)
-        .show()
-        .map_err(|e| e.to_string())
-}
-
 fn memory_file_path(memory: &MemoryManager, name: &str) -> Result<std::path::PathBuf, String> {
     let name = name.trim();
     if name.is_empty() || name.contains(['/', '\\']) || name.contains("..") {
@@ -3942,7 +3912,9 @@ async fn send_message_inner(
             .await
             .map_err(|error| error.to_string())?;
         if let Some(owner_id) = owner.filter(|owner_id| owner_id != &ap.id) {
-            ap = load_active_project(&state, &owner_id).await?.0;
+            ap = project_commands::load_active_project(&state, &owner_id)
+                .await?
+                .0;
         }
     }
     let _project_activity = state.begin_project_activity(&ap.id)?;
@@ -5067,7 +5039,11 @@ async fn generate_review(
             .await
             .map_err(|error| error.to_string())?
             .ok_or_else(|| "Session project was not found.".to_string())?;
-        Some(load_active_project(state, &project_id).await?.0)
+        Some(
+            project_commands::load_active_project(state, &project_id)
+                .await?
+                .0,
+        )
     } else {
         None
     };
@@ -5513,1701 +5489,6 @@ async fn side_chat(
     Ok(completion.content)
 }
 
-#[tauri::command]
-async fn new_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<String, String> {
-    // Create a fresh frame and hand its id to the UI up front, so the UI can
-    // route streamed events to the right transcript *before* the first delta
-    // arrives. Does NOT cancel any running turn — parallel conversations keep
-    // running. Empty frames are filtered out of the sidebar until they get a
-    // user message.
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let id = create_session_frame(&state.store, &ap.id).await?;
-    state.set_active_frame(window.label(), Some(id.clone()));
-    Ok(id)
-}
-
-#[tauri::command]
-async fn branch_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    session_id: Option<String>,
-    title: Option<String>,
-    user_index: Option<usize>,
-) -> Result<String, String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let id = create_session_frame(&state.store, &ap.id).await?;
-    if let Some(source) = session_id.as_deref().filter(|s| !s.is_empty()) {
-        let model_id = models::session_profile_id(&state.store, source).await;
-        state
-            .store
-            .set_frame_model(&id, &ap.id, &model_id)
-            .await
-            .map_err(|error| error.to_string())?;
-        let msgs = state
-            .store
-            .load_messages(source)
-            .await
-            .map_err(|e| format!("{e}"))?;
-        let keep = user_index
-            .map(|idx| user_message_start(&msgs, idx))
-            .unwrap_or(msgs.len());
-        for (idx, msg) in msgs.iter().take(keep).enumerate() {
-            state
-                .store
-                .append_message(&id, idx as i64 + 1, msg)
-                .await
-                .map_err(|e| format!("{e}"))?;
-        }
-    }
-    if let Some(t) = branch_title(title.as_deref()) {
-        let _ = state.store.rename_session(&id, &ap.id, &t).await;
-    }
-    state.set_active_frame(window.label(), Some(id.clone()));
-    Ok(id)
-}
-
-#[tauri::command]
-async fn list_sessions(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<Vec<SessionInfo>, String> {
-    let ap = state.active(window.label());
-    let rows = state
-        .store
-        .list_sessions(&ap.id)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    let pinned_ids = state
-        .store
-        .list_pinned_sessions(&ap.id)
-        .await
-        .map(|rows| rows.into_iter().map(|row| row.0).collect::<HashSet<_>>())
-        .unwrap_or_default();
-    let running = state.running_turns.lock().await.clone();
-    Ok(rows
-        .into_iter()
-        .map(|(id, title, ts, folder_id)| SessionInfo {
-            running: running.contains(&id),
-            pinned: pinned_ids.contains(&id),
-            id,
-            title,
-            ts,
-            folder_id,
-        })
-        .collect())
-}
-
-#[tauri::command]
-async fn list_sessions_page(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    cursor: Option<SessionCursor>,
-) -> Result<SessionPage, String> {
-    let ap = state.active(window.label());
-    let mut rows = state
-        .store
-        .list_sessions_page(
-            &ap.id,
-            cursor
-                .as_ref()
-                .map(|cursor| (cursor.ts, cursor.id.as_str())),
-            SESSION_HISTORY_PAGE_SIZE + 1,
-        )
-        .await
-        .map_err(|e| format!("{e}"))?;
-    let has_more = rows.len() > SESSION_HISTORY_PAGE_SIZE;
-    rows.truncate(SESSION_HISTORY_PAGE_SIZE);
-    let next_cursor = has_more.then(|| {
-        let row = rows.last().expect("a full session page has a final row");
-        SessionCursor {
-            ts: row.2,
-            id: row.0.clone(),
-        }
-    });
-    // Pinned sessions float to the top and must show even when they fall outside
-    // the newest keyset page, so fetch them once (first page only) and prepend any
-    // that aren't already in this page. The keyset cursor is left untouched.
-    let pinned_rows = match cursor {
-        None => state
-            .store
-            .list_pinned_sessions(&ap.id)
-            .await
-            .map_err(|e| format!("{e}"))?,
-        Some(_) => Vec::new(),
-    };
-    let pinned_ids: HashSet<String> = pinned_rows.iter().map(|row| row.0.clone()).collect();
-    let page_ids: HashSet<String> = rows.iter().map(|row| row.0.clone()).collect();
-    let running = state.running_turns.lock().await.clone();
-    let mut items: Vec<SessionInfo> = pinned_rows
-        .into_iter()
-        .filter(|(id, ..)| !page_ids.contains(id))
-        .map(|(id, title, ts, folder_id)| SessionInfo {
-            running: running.contains(&id),
-            pinned: true,
-            id,
-            title,
-            ts,
-            folder_id,
-        })
-        .collect();
-    items.extend(
-        rows.into_iter()
-            .map(|(id, title, ts, folder_id)| SessionInfo {
-                running: running.contains(&id),
-                pinned: pinned_ids.contains(&id),
-                id,
-                title,
-                ts,
-                folder_id,
-            }),
-    );
-    Ok(SessionPage {
-        items,
-        next_cursor,
-        running_ids: running.into_iter().collect(),
-    })
-}
-
-#[tauri::command]
-async fn list_execution_contexts(
-    state: State<'_, AppState>,
-) -> Result<Vec<wisp_store::ExecutionContext>, String> {
-    state
-        .store
-        .list_execution_contexts()
-        .await
-        .map_err(|e| format!("{e}"))
-}
-
-#[tauri::command]
-fn list_runtimes(state: State<'_, AppState>) -> Vec<wisp_runtime::RuntimeInfo> {
-    state.runtime_manager.list()
-}
-
-#[tauri::command]
-async fn inspect_runtime(
-    state: State<'_, AppState>,
-    project_id: String,
-    context_id: String,
-    language: wisp_runtime::RuntimeLanguage,
-) -> Result<wisp_runtime::RuntimeObjectList, String> {
-    state
-        .runtime_manager
-        .inspect(&wisp_runtime::RuntimeKey {
-            project_id,
-            context_id,
-            language,
-        })
-        .await
-        .map_err(|error| error.to_string())
-}
-
-/// Run code the user selected in the file preview against their bound runtime.
-/// Deferred in the runtime design until the UI gained a code editor; it has one
-/// now. The user is looking at the code they pressed Run on, so this path is
-/// deliberately outside the agent tool-approval flow.
-///
-/// Returns console text. Code that raised is still `Ok`: `format_response` tags
-/// it `[error]` exactly as the agent tools render it. `Err` means the runtime
-/// itself never produced a result.
-#[tauri::command]
-async fn execute_runtime(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    context_id: String,
-    language: wisp_runtime::RuntimeLanguage,
-    code: String,
-) -> Result<String, String> {
-    if code.len() > wisp_runtime::MAX_CODE_BYTES {
-        return Err(format!(
-            "Selection exceeds the {} byte runtime limit.",
-            wisp_runtime::MAX_CODE_BYTES
-        ));
-    }
-    let project = state.active(window.label());
-    let key = wisp_runtime::RuntimeKey {
-        project_id: project.id,
-        context_id,
-        language,
-    };
-    let mut execution = state
-        .runtime_manager
-        .execute(&key, &project.root, code)
-        .await
-        .map_err(|error| error.to_string())?;
-    loop {
-        match execution.recv().await {
-            // ponytail: buffered, not streamed — the final frame repeats every
-            // chunk. Stream to the console when a cell runs long enough to care.
-            Some(wisp_runtime::RuntimeEvent::Stdout(_)) => {}
-            Some(wisp_runtime::RuntimeEvent::Finished(Ok(response))) => {
-                return Ok(wisp_runtime::format_response(&response))
-            }
-            Some(wisp_runtime::RuntimeEvent::Finished(Err(error))) => return Err(error),
-            None => return Err("Runtime ended before returning a result.".into()),
-        }
-    }
-}
-
-#[tauri::command]
-async fn start_runtime(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    context_id: String,
-    language: wisp_runtime::RuntimeLanguage,
-) -> Result<wisp_runtime::RuntimeInfo, String> {
-    let project = state.active(window.label());
-    state
-        .runtime_manager
-        .start(
-            wisp_runtime::RuntimeKey {
-                project_id: project.id,
-                context_id,
-                language,
-            },
-            project.root,
-        )
-        .await
-        .map_err(|error| error.to_string())
-}
-
-#[tauri::command]
-async fn stop_runtime(
-    state: State<'_, AppState>,
-    project_id: String,
-    context_id: String,
-    language: wisp_runtime::RuntimeLanguage,
-) -> Result<Option<wisp_runtime::RuntimeInfo>, String> {
-    Ok(state
-        .runtime_manager
-        .stop(&wisp_runtime::RuntimeKey {
-            project_id,
-            context_id,
-            language,
-        })
-        .await)
-}
-
-#[tauri::command]
-async fn restart_runtime(
-    state: State<'_, AppState>,
-    project_id: String,
-    context_id: String,
-    language: wisp_runtime::RuntimeLanguage,
-) -> Result<wisp_runtime::RuntimeInfo, String> {
-    let (_, workspace) = state
-        .store
-        .get_project(&project_id)
-        .await
-        .map_err(|error| error.to_string())?
-        .ok_or_else(|| format!("Project not found: {project_id}"))?;
-    let root = ensure_writable(PathBuf::from(workspace), &state.app_data);
-    state
-        .runtime_manager
-        .restart(
-            wisp_runtime::RuntimeKey {
-                project_id,
-                context_id,
-                language,
-            },
-            root,
-        )
-        .await
-        .map_err(|error| error.to_string())
-}
-
-#[tauri::command]
-async fn list_runs(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<Vec<wisp_store::RunRecord>, String> {
-    let ap = state.active(window.label());
-    state
-        .store
-        .list_runs_by_project(&ap.id)
-        .await
-        .map_err(|e| format!("{e}"))
-}
-
-#[tauri::command]
-async fn get_run(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    run_id: String,
-) -> Result<wisp_store::RunRecord, String> {
-    let ap = state.active(window.label());
-    let run = state
-        .store
-        .get_run(&run_id)
-        .await
-        .map_err(|e| e.to_string())?
-        .ok_or_else(|| "Run not found".to_string())?;
-    if run.project_id != ap.id {
-        return Err("Run does not belong to the active project".into());
-    }
-    Ok(run)
-}
-
-#[tauri::command]
-async fn cancel_run(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    run_id: String,
-) -> Result<wisp_store::RunRecord, String> {
-    let ap = state.active(window.label());
-    let run = state
-        .store
-        .get_run(&run_id)
-        .await
-        .map_err(|e| e.to_string())?
-        .ok_or_else(|| "Run not found".to_string())?;
-    if run.project_id != ap.id {
-        return Err("Run does not belong to the active project".into());
-    }
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    state.run_manager.cancel(&state.store, &run_id).await?;
-    state
-        .store
-        .get_run(&run_id)
-        .await
-        .map_err(|e| e.to_string())?
-        .ok_or_else(|| "Run disappeared after cancellation".to_string())
-}
-
-#[tauri::command]
-async fn get_research_graph(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<wisp_store::ResearchGraph, String> {
-    let ap = state.active(window.label());
-    state
-        .store
-        .research_graph(&ap.id)
-        .await
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-async fn list_folders(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<Vec<FolderInfo>, String> {
-    let ap = state.active(window.label());
-    let rows = state
-        .store
-        .list_folders(&ap.id)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(rows
-        .into_iter()
-        .map(|(id, name, _)| FolderInfo { id, name })
-        .collect())
-}
-
-#[tauri::command]
-async fn create_folder(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    name: String,
-) -> Result<FolderInfo, String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let id = Uuid::new_v4().to_string();
-    state
-        .store
-        .create_folder(&id, &ap.id, &name)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(FolderInfo {
-        id,
-        name: name.trim().to_string(),
-    })
-}
-
-#[tauri::command]
-async fn rename_folder(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-    name: String,
-) -> Result<(), String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    state
-        .store
-        .rename_folder(&id, &ap.id, &name)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(())
-}
-
-#[tauri::command]
-async fn delete_folder(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-) -> Result<(), String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    state
-        .store
-        .delete_folder(&id, &ap.id)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(())
-}
-
-#[tauri::command]
-async fn move_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-    folder_id: Option<String>,
-) -> Result<(), String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    state
-        .store
-        .move_session_to_folder(&id, &ap.id, folder_id.as_deref())
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(())
-}
-
-#[tauri::command]
-async fn transfer_session_to_project(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-    target_project_id: String,
-    mode: String,
-) -> Result<String, String> {
-    let source = state.active(window.label());
-    if target_project_id == source.id {
-        return Err("Source and target projects must be different.".into());
-    }
-    if state
-        .store
-        .get_project(&target_project_id)
-        .await
-        .map_err(|error| error.to_string())?
-        .is_none()
-    {
-        return Err("Target project not found.".into());
-    }
-    let owner = state
-        .store
-        .frame_project_id(&id)
-        .await
-        .map_err(|error| error.to_string())?;
-    if owner.as_deref() != Some(source.id.as_str()) {
-        return Err("Session does not belong to the active project.".into());
-    }
-    let remove_source = match mode.as_str() {
-        "copy" => false,
-        "move" => true,
-        _ => return Err("Transfer mode must be 'copy' or 'move'.".into()),
-    };
-
-    let session_is_busy = || {
-        state.awaiting_confirm.lock().unwrap().contains(&id)
-            || state.reviewing.lock().unwrap().contains(&id)
-    };
-    if state.running_turns.lock().await.contains(&id) || session_is_busy() {
-        return Err(
-            "Wait for the session to finish its turn, approval, or review before transferring it."
-                .into(),
-        );
-    }
-
-    let _source_activity = state.begin_project_activity(&source.id)?;
-    let _target_activity = state.begin_project_activity(&target_project_id)?;
-    let runtime = state.sessions.lock().await.get(&id).cloned();
-    let _workflow_guard = match runtime.as_ref() {
-        Some(runtime) => Some(runtime.workflow.lock().await),
-        None => None,
-    };
-    let _agent_guard = match runtime.as_ref() {
-        Some(runtime) => Some(runtime.agent.lock().await),
-        None => None,
-    };
-    if state.running_turns.lock().await.contains(&id) || session_is_busy() {
-        return Err(
-            "Wait for the session to finish its turn, approval, or review before transferring it."
-                .into(),
-        );
-    }
-
-    let new_id = Uuid::new_v4().to_string();
-    if remove_source {
-        state
-            .store
-            .move_session_to_project(&id, &source.id, &target_project_id, &new_id)
-            .await
-            .map_err(|error| error.to_string())?;
-        if let Some(runtime) = runtime.as_ref() {
-            runtime.deleted.store(true, Ordering::SeqCst);
-            runtime.cancel.store(true, Ordering::Relaxed);
-        }
-        acp::close_frame(&state, &id).await;
-        state.sessions.lock().await.remove(&id);
-        if state.active_frame(window.label()).as_deref() == Some(id.as_str()) {
-            state.set_active_frame(window.label(), None);
-        }
-    } else {
-        state
-            .store
-            .copy_session_to_project(&id, &source.id, &target_project_id, &new_id)
-            .await
-            .map_err(|error| error.to_string())?;
-    }
-    Ok(new_id)
-}
-
-#[tauri::command]
-async fn delete_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-) -> Result<(), String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let owner = state
-        .store
-        .frame_project_id(&id)
-        .await
-        .map_err(|error| error.to_string())?;
-    if owner.as_deref() != Some(ap.id.as_str()) {
-        return Err("Session does not belong to the active project.".into());
-    }
-    let runtime = state.sessions.lock().await.get(&id).cloned();
-    if let Some(rt) = runtime.as_ref() {
-        rt.deleted.store(true, Ordering::SeqCst);
-        rt.cancel.store(true, Ordering::Relaxed);
-    }
-    acp::cancel_frame(&state, &id).await;
-    // Match send/Plan lock order. The tombstone prevents work already queued
-    // behind these guards from restarting after the DB cascade.
-    let _workflow_guard = match runtime.as_ref() {
-        Some(rt) => Some(rt.workflow.lock().await),
-        None => None,
-    };
-    let _agent_guard = match runtime.as_ref() {
-        Some(rt) => Some(rt.agent.lock().await),
-        None => None,
-    };
-    acp::close_frame(&state, &id).await;
-    state.sessions.lock().await.remove(&id);
-    if state.active_frame(window.label()).as_deref() == Some(id.as_str()) {
-        state.set_active_frame(window.label(), None);
-    }
-    state
-        .store
-        .delete_session(&id, &ap.id)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(())
-}
-
-#[tauri::command]
-async fn rename_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-    title: String,
-) -> Result<(), String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    state
-        .store
-        .rename_session(&id, &ap.id, &title)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(())
-}
-
-#[tauri::command]
-async fn set_session_pinned(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-    pinned: bool,
-) -> Result<(), String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    state
-        .store
-        .set_session_pinned(&id, &ap.id, pinned)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(())
-}
-
-/// How many sessions appear on the Projects landing "Recent sessions" column.
-const RECENT_SESSIONS_LIMIT: i64 = 5;
-
-#[tauri::command]
-async fn list_recent_sessions(
-    state: State<'_, AppState>,
-) -> Result<Vec<serde_json::Value>, String> {
-    let running = state.running_turns.lock().await.clone();
-    let awaiting = state.awaiting_confirm.lock().unwrap().clone();
-    let rows = state
-        .store
-        .list_recent_sessions_detail(RECENT_SESSIONS_LIMIT)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(rows
-        .into_iter()
-        .map(|r| {
-            let status = session_runtime_status(
-                &r.id,
-                r.last_role.as_deref(),
-                r.unseen,
-                &running,
-                &awaiting,
-            );
-            serde_json::json!({
-                "id": r.id,
-                "project_id": r.project_id,
-                "title": r.title,
-                "ts": r.created_at,
-                "status": status,
-            })
-        })
-        .collect())
-}
-
-#[tauri::command]
-async fn list_projects(state: State<'_, AppState>) -> Result<Vec<ProjectSummary>, String> {
-    let running = state.running_turns.lock().await.clone();
-    let awaiting = state.awaiting_confirm.lock().unwrap().clone();
-    let rows = state
-        .store
-        .list_projects()
-        .await
-        .map_err(|e| format!("{e}"))?;
-    let mut out = vec![];
-    for (id, name, ws, _c, upd, cnt, desc) in rows {
-        let (running_count, needs_you_count) =
-            project_status_counts(&state.store, &id, &running, &awaiting).await;
-        let sync_state = state.store.get_project_sync_state(&id).await.ok().flatten();
-        let sync_configured = sync_state
-            .as_ref()
-            .is_some_and(|state| state.base_revision.is_some());
-        out.push(ProjectSummary {
-            id,
-            name,
-            description: desc,
-            workspace_dir: ws,
-            session_count: cnt,
-            updated_at: upd,
-            running_count,
-            needs_you_count,
-            sync_configured,
-            last_synced_at: sync_state.and_then(|state| state.last_synced_at),
-        });
-    }
-    Ok(out)
-}
-
-#[tauri::command]
-async fn pick_directory(app: AppHandle) -> Result<Option<String>, String> {
-    use tauri_plugin_dialog::DialogExt;
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    app.dialog().file().pick_folder(move |p| {
-        let _ = tx.send(p);
-    });
-    let picked = rx.await.map_err(|e| format!("{e}"))?;
-    Ok(picked.map(|fp| fp.to_string()))
-}
-
-/// Save UI-provided text via the native save dialog. Returns the saved path,
-/// or `None` if the user cancelled.
-#[tauri::command]
-async fn export_text_file(
-    app: AppHandle,
-    default_name: String,
-    content: String,
-) -> Result<Option<String>, String> {
-    use tauri_plugin_dialog::DialogExt;
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    app.dialog()
-        .file()
-        .set_file_name(&default_name)
-        .save_file(move |p| {
-            let _ = tx.send(p);
-        });
-    let Some(dest) = rx.await.map_err(|e| format!("{e}"))? else {
-        return Ok(None);
-    };
-    let dest = std::path::PathBuf::from(dest.to_string());
-    tokio::fs::write(&dest, content)
-        .await
-        .map_err(|e| format!("write failed: {e}"))?;
-    Ok(Some(dest.to_string_lossy().into_owned()))
-}
-
-/// Pick a JSON file via the native open dialog and return its content, or
-/// `None` if the user cancelled.
-#[tauri::command]
-async fn import_json_file(app: AppHandle) -> Result<Option<String>, String> {
-    use tauri_plugin_dialog::DialogExt;
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    app.dialog()
-        .file()
-        .add_filter("JSON", &["json"])
-        .pick_file(move |p| {
-            let _ = tx.send(p);
-        });
-    let Some(path) = rx.await.map_err(|e| format!("{e}"))? else {
-        return Ok(None);
-    };
-    let content = tokio::fs::read_to_string(std::path::PathBuf::from(path.to_string()))
-        .await
-        .map_err(|e| format!("read failed: {e}"))?;
-    Ok(Some(content))
-}
-
-/// Copy a workspace file to a user-chosen location via the native save dialog.
-/// Returns the saved path, or `None` if the user cancelled.
-fn parse_ssh_artifact_uri(uri: &str) -> Option<(String, String)> {
-    let rest = uri.strip_prefix("ssh://")?;
-    let (context, path) = rest.split_once('/')?;
-    if context.is_empty() || path.is_empty() {
-        return None;
-    }
-    let remote_path = if path.starts_with("~/") {
-        path.to_string()
-    } else {
-        format!("/{path}")
-    };
-    Some((format!("ssh:{context}"), remote_path))
-}
-
-#[tauri::command]
-async fn download_file(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    path: String,
-) -> Result<Option<String>, String> {
-    use tauri_plugin_dialog::DialogExt;
-    let ap = state.active(window.label());
-    let remote = parse_ssh_artifact_uri(&path);
-    let local = if remote.is_none() {
-        let real = wisp_tools::safety::validate_file_path(&ap.root, &path)?;
-        if !real.is_file() {
-            return Err(format!("file not found: {path}"));
-        }
-        Some(real)
-    } else {
-        None
-    };
-    let default_name = std::path::Path::new(
-        remote
-            .as_ref()
-            .map(|(_, path)| path.as_str())
-            .unwrap_or_else(|| local.as_ref().unwrap().to_str().unwrap_or("download")),
-    )
-    .file_name()
-    .and_then(|n| n.to_str())
-    .unwrap_or("download")
-    .to_string();
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    app.dialog()
-        .file()
-        .set_file_name(&default_name)
-        .save_file(move |p| {
-            let _ = tx.send(p);
-        });
-    let Some(dest) = rx.await.map_err(|e| format!("{e}"))? else {
-        return Ok(None); // user cancelled
-    };
-    let dest_path = std::path::PathBuf::from(dest.to_string());
-    if let Some((context_id, remote_path)) = remote {
-        let frame_id = state.active_frame(window.label());
-        let context = state
-            .store
-            .get_execution_context(&context_id)
-            .await
-            .map_err(|e| format!("{e}"))?
-            .ok_or_else(|| format!("SSH execution context not found: {context_id}"))?;
-        state
-            .run_manager
-            .download_ssh_file(
-                &state.store,
-                &ap.id,
-                frame_id.as_deref(),
-                &context,
-                &remote_path,
-                &dest_path,
-            )
-            .await?;
-    } else {
-        tokio::fs::copy(local.unwrap(), &dest_path)
-            .await
-            .map_err(|e| format!("copy failed: {e}"))?;
-    }
-    Ok(Some(dest_path.to_string_lossy().into_owned()))
-}
-
-#[tauri::command]
-async fn create_project(
-    state: State<'_, AppState>,
-    name: String,
-    workspace_dir: String,
-    description: String,
-    agent_context: String,
-    standard_layout: bool,
-) -> Result<ProjectSummary, String> {
-    if name.trim().is_empty() {
-        return Err("Project name is required".into());
-    }
-    let dir = workspace_dir.trim();
-    if dir.is_empty() {
-        return Err("A working directory is required".into());
-    }
-    let path = PathBuf::from(dir);
-    std::fs::create_dir_all(&path)
-        .map_err(|e| format!("Failed to create working directory: {e}"))?;
-    // Writability probe: create + remove a temp marker.
-    let marker = path.join(".wisp-write-test");
-    std::fs::write(&marker, b"").map_err(|e| format!("Working directory is not writable: {e}"))?;
-    let _ = std::fs::remove_file(&marker);
-
-    let id = Uuid::new_v4().to_string();
-    // #405: opt-in. Unchecked means the user keeps their own structure, so we
-    // create nothing — the convention lives in .wisp/WISP.md instead (below).
-    if standard_layout {
-        workspace_manifest::init_workspace_layout(&path, &id, name.trim())?;
-    }
-    state
-        .store
-        .create_project(&id, name.trim(), dir)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    // Description (DB) + Agent Context (.wisp/WISP.md) — same storage as update_project.
-    let desc = description.trim();
-    if !desc.is_empty() {
-        state
-            .store
-            .update_project(&id, name.trim(), desc)
-            .await
-            .map_err(|e| format!("{e}"))?;
-    }
-    let ctx = agent_context.trim();
-    if !ctx.is_empty() {
-        let wisp_dir = path.join(".wisp");
-        std::fs::create_dir_all(&wisp_dir)
-            .map_err(|e| format!("Failed to write Agent Context: {e}"))?;
-        std::fs::write(wisp_dir.join("WISP.md"), ctx)
-            .map_err(|e| format!("Failed to write Agent Context: {e}"))?;
-    }
-    Ok(build_project_summary(&state, &id).await)
-}
-
-/// Cancel and drop every in-memory runtime belonging to `project_id`'s sessions
-/// (e.g. the project is being deleted). Other projects' sessions keep running —
-/// switching/closing a project must not stop unrelated work (#52). Call this
-/// *before* the project's frames are removed from the store.
-async fn cancel_project_sessions(state: &AppState, project_id: &str) {
-    let frame_ids: Vec<String> = state
-        .store
-        .list_sessions(project_id)
-        .await
-        .map(|rows| rows.into_iter().map(|(id, ..)| id).collect())
-        .unwrap_or_default();
-    let runtimes = {
-        let sessions = state.sessions.lock().await;
-        frame_ids
-            .iter()
-            .filter_map(|fid| sessions.get(fid).cloned().map(|rt| (fid.clone(), rt)))
-            .collect::<Vec<_>>()
-    };
-    for (_, rt) in &runtimes {
-        rt.deleted.store(true, Ordering::SeqCst);
-        rt.cancel.store(true, Ordering::SeqCst);
-    }
-    for fid in &frame_ids {
-        acp::cancel_frame(state, fid).await;
-    }
-    for (_, rt) in &runtimes {
-        let _workflow = rt.workflow.lock().await;
-        let _agent = rt.agent.lock().await;
-    }
-    for fid in &frame_ids {
-        acp::close_frame(state, fid).await;
-    }
-    {
-        let mut sessions = state.sessions.lock().await;
-        for fid in &frame_ids {
-            sessions.remove(fid);
-        }
-    }
-    let mut running = state.running_turns.lock().await;
-    for fid in &frame_ids {
-        running.remove(fid);
-    }
-}
-
-/// Point the backend's active project at `id`, rebuilding its skills/memory.
-/// Returns the resolved `(name, workspace_dir)`. `id` must exist in the store.
-///
-/// Switching projects no longer tears down the previous project's sessions —
-/// each session's agent already captured its own root/skills/memory at creation,
-/// so cross-project turns run in parallel and stay monitorable on the dashboard
-/// (#52). Deleting a project stops only *its* sessions (see `delete_project`).
-/// Build a project's ActiveProject bundle (root, skills, memory) by id, plus
-/// its (name, workspace) for callers that need them. Pure load — does not touch
-/// the per-window active slot.
-async fn load_active_project(
-    state: &AppState,
-    id: &str,
-) -> Result<(ActiveProject, String, String), String> {
-    let (name, ws) = state
-        .store
-        .get_project(id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| "Project not found".to_string())?;
-    let root = ensure_writable(PathBuf::from(&ws), &state.app_data);
-    let skills = Arc::new(SkillIndex::load(&skill_paths(&root)));
-    let memory = Arc::new(MemoryManager::new(&root));
-    Ok((
-        ActiveProject {
-            id: id.to_string(),
-            root,
-            skills,
-            memory,
-        },
-        name,
-        ws,
-    ))
-}
-
-async fn set_active_project(
-    state: &AppState,
-    label: &str,
-    id: &str,
-) -> Result<(String, String), String> {
-    let (ap, name, ws) = load_active_project(state, id).await?;
-    let root = ap.root.clone();
-    state.set_active(label, ap);
-    state.set_active_frame(label, None);
-    {
-        state.bootstrap.lock().unwrap().workspace = root.to_string_lossy().into_owned();
-    }
-    let _ = state.store.set_setting("active_project_id", id).await;
-    Ok((name, ws))
-}
-
-#[tauri::command]
-async fn open_project(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-) -> Result<ProjectSummary, String> {
-    let _project_activity = state.begin_project_activity(&id)?;
-    let (name, ws) = set_active_project(state.inner(), window.label(), &id).await?;
-    let _ = state.store.create_project(&id, &name, &ws).await; // touch updated_at → sorts to top
-    Ok(build_project_summary(&state, &id).await)
-}
-
-/// Project ids that currently have their own window, persisted so the set can be
-/// restored on the next launch (#52, Phase 3). Stored as a JSON array setting.
-async fn persisted_windows(store: &Store) -> Vec<String> {
-    store
-        .get_setting("open_project_windows")
-        .await
-        .ok()
-        .flatten()
-        .and_then(|s| serde_json::from_str::<Vec<String>>(&s).ok())
-        .unwrap_or_default()
-}
-
-async fn update_persisted_windows(store: &Store, id: &str, present: bool) {
-    let mut v = persisted_windows(store).await;
-    let had = v.iter().any(|x| x == id);
-    if present && !had {
-        v.push(id.to_string());
-    } else if !present && had {
-        v.retain(|x| x != id);
-    } else {
-        return;
-    }
-    let _ = store
-        .set_setting(
-            "open_project_windows",
-            &serde_json::to_string(&v).unwrap_or_default(),
-        )
-        .await;
-}
-
-fn project_window_label(id: &str) -> String {
-    format!("proj-{id}") // project ids are UUIDs or "default" — label-safe
-}
-
-/// URL for a dedicated project window. Ids are UUIDs or "default" — no
-/// percent-encoding needed (matches `url_project_param` in the frontend).
-fn project_window_url(id: &str, session: Option<&str>) -> String {
-    match session {
-        Some(sid) => format!("index.html?project={id}&session={sid}"),
-        None => format!("index.html?project={id}"),
-    }
-}
-
-/// Open a project in its own window (or focus the existing one), wiring up
-/// cleanup on close. Shared by the `open_project_window` command and the
-/// startup restore (#52). With `session`, the window opens straight into that
-/// session — an existing window is told via the `open-session` event (#423).
-async fn spawn_project_window(
-    app: &AppHandle,
-    state: &AppState,
-    id: &str,
-    session: Option<&str>,
-) -> Result<String, String> {
-    let label = project_window_label(id);
-    if let Some(w) = app.get_webview_window(&label) {
-        let _ = w.set_focus();
-        if let Some(sid) = session {
-            let _ = app.emit_to(
-                label.as_str(),
-                "open-session",
-                serde_json::json!({ "projectId": id, "sessionId": sid }),
-            );
-        }
-        return Ok(label);
-    }
-    // Pre-set this window's active project so its first commands resolve correctly
-    // even before the window's frontend calls open_project.
-    set_active_project(state, &label, id).await?;
-    let url = tauri::WebviewUrl::App(project_window_url(id, session).into());
-    let builder = tauri::WebviewWindowBuilder::new(app, &label, url)
-        .title("wisp-science")
-        .inner_size(1100.0, 760.0)
-        .resizable(true);
-    #[cfg(target_os = "windows")]
-    let builder = builder.decorations(false).shadow(true);
-    let win = builder.build().map_err(|e| e.to_string())?;
-    #[cfg(target_os = "macos")]
-    wire_macos_menu_events(&win);
-    let evt_app = app.clone();
-    let evt_label = label.clone();
-    let evt_id = id.to_string();
-    win.on_window_event(move |ev| {
-        if matches!(ev, tauri::WindowEvent::Destroyed) {
-            // Drop this window's per-window project context and stop persisting
-            // it for restore. Its running sessions are tracked globally and keep
-            // going until they finish or are stopped.
-            let st = evt_app.state::<AppState>();
-            st.active.write().unwrap().remove(&evt_label);
-            st.active_frame.write().unwrap().remove(&evt_label);
-            let store = st.store.clone();
-            let id = evt_id.clone();
-            tauri::async_runtime::spawn(async move {
-                update_persisted_windows(&store, &id, false).await;
-            });
-        }
-    });
-    update_persisted_windows(&state.store, id, true).await;
-    Ok(label)
-}
-
-/// Open a project in its own window (or focus the existing one). Each window
-/// carries its own active project, keyed by window label (#52).
-#[tauri::command]
-async fn open_project_window(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    id: String,
-    session: Option<String>,
-) -> Result<String, String> {
-    spawn_project_window(&app, state.inner(), &id, session.as_deref()).await
-}
-
-#[tauri::command]
-async fn delete_project(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-) -> Result<(), String> {
-    let _project_activity = state.begin_project_activity(&id)?;
-    // The delete ✕ is only reachable from the projects list, so a project may
-    // legitimately be deleted while it's still the backend's *active* one
-    // (returning to the list is a frontend-only nav — it never told the backend
-    // to leave). Delete it, then fall back to the always-present "default"
-    // workspace so `active` never dangles at a deleted project.
-    let was_active = state.active(window.label()).id == id;
-    // Stop the deleted project's own running sessions (gather frame ids before
-    // the store cascade removes them); other projects keep running (#52).
-    cancel_project_sessions(state.inner(), &id).await;
-    state.runtime_manager.stop_project(&id).await;
-    state
-        .store
-        .delete_project(&id)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    project_sync::forget_project_key(&id).await;
-    if was_active {
-        let _ = set_active_project(state.inner(), window.label(), "default").await;
-    }
-    Ok(())
-}
-
-#[derive(Serialize, Clone)]
-struct ProjectSettings {
-    id: String,
-    name: String,
-    description: String,
-    agent_context: String,
-}
-
-/// Read the active project's editable settings for the Project Settings modal.
-/// Agent Context is `.wisp/WISP.md`, injected into every seeded system prompt.
-#[tauri::command]
-async fn get_project_settings(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<ProjectSettings, String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let (name, description, _ws) = state
-        .store
-        .get_project_meta(&ap.id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .unwrap_or_default();
-    let agent_context =
-        std::fs::read_to_string(ap.root.join(".wisp").join("WISP.md")).unwrap_or_default();
-    Ok(ProjectSettings {
-        id: ap.id.clone(),
-        name,
-        description,
-        agent_context,
-    })
-}
-
-/// Save the active project's name/description (DB) and Agent Context (.wisp/WISP.md).
-/// An empty Agent Context removes WISP.md so the prompt falls back to "no rules".
-/// Takes effect on the next seeded session; already-running agents keep their prompt.
-#[tauri::command]
-async fn update_project(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    name: String,
-    description: String,
-    agent_context: String,
-) -> Result<ProjectSummary, String> {
-    if name.trim().is_empty() {
-        return Err("Project name is required".into());
-    }
-    let ap = state.active(window.label());
-    state
-        .store
-        .update_project(&ap.id, name.trim(), description.trim())
-        .await
-        .map_err(|e| format!("{e}"))?;
-    let wisp_dir = ap.root.join(".wisp");
-    let wisp_md = wisp_dir.join("WISP.md");
-    let ctx = agent_context.trim();
-    if ctx.is_empty() {
-        let _ = std::fs::remove_file(&wisp_md);
-    } else {
-        std::fs::create_dir_all(&wisp_dir)
-            .map_err(|e| format!("Failed to write Agent Context: {e}"))?;
-        std::fs::write(&wisp_md, ctx).map_err(|e| format!("Failed to write Agent Context: {e}"))?;
-    }
-    Ok(build_project_summary(&state, &ap.id).await)
-}
-
-/// Switch the active session to `id`, load its transcript, and return the
-/// rendered rows so the UI can repopulate the conversation view.
-/// Rewind the named session to just before the given user turn (for message
-/// edit). Only touches that session's agent context and DB rows.
-#[tauri::command]
-async fn rewind_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    session_id: Option<String>,
-    user_index: usize,
-) -> Result<(), String> {
-    let frame_id = match session_id.as_deref().filter(|s| !s.is_empty()) {
-        Some(id) => id.to_string(),
-        None => state
-            .active_frame(window.label())
-            .ok_or_else(|| "No active session to rewind.".to_string())?,
-    };
-    let project_id = state
-        .store
-        .frame_project_id(&frame_id)
-        .await
-        .map_err(|error| error.to_string())?
-        .ok_or_else(|| "Session project was not found.".to_string())?;
-    let _project_activity = state.begin_project_activity(&project_id)?;
-    if state
-        .store
-        .get_acp_session(&frame_id)
-        .await
-        .map_err(|error| error.to_string())?
-        .is_some()
-    {
-        return Err("ACP sessions cannot be rewound in protocol v1.".into());
-    }
-    let rt = state.sessions.lock().await.get(&frame_id).cloned();
-    let keep = if let Some(rt) = rt {
-        let mut guard = rt.agent.lock().await;
-        if let Some(agent) = guard.as_mut() {
-            let k = user_message_start(&agent.ctx.messages, user_index);
-            agent.ctx.messages.truncate(k);
-            k
-        } else {
-            user_index_to_keep_after_db(&state.store, &frame_id, user_index).await?
-        }
-    } else {
-        user_index_to_keep_after_db(&state.store, &frame_id, user_index).await?
-    };
-    state
-        .store
-        .truncate_messages(&frame_id, keep as i64)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    if let Some(rt) = state.sessions.lock().await.get(&frame_id) {
-        rt.set_last_seq(keep as i64);
-    }
-    Ok(())
-}
-
-/// Compute the `keep` index purely from persisted messages when no in-memory
-/// agent exists for the session yet.
-async fn user_index_to_keep_after_db(
-    store: &Store,
-    frame_id: &str,
-    user_index: usize,
-) -> Result<usize, String> {
-    let msgs = store
-        .load_messages(frame_id)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(user_message_start(&msgs, user_index))
-}
-
-fn transcript_page_items(page: &wisp_store::SessionTranscriptPage) -> Result<Vec<UiItem>, String> {
-    let msgs = page
-        .messages
-        .iter()
-        .map(|(_, message)| message.clone())
-        .collect::<Vec<_>>();
-    let events: Vec<AgentEvent> = page
-        .ui_events
-        .iter()
-        .map(|json| serde_json::from_str(json))
-        .collect::<Result<_, _>>()
-        .map_err(|e| format!("invalid persisted UI event: {e}"))?;
-    let (mut items, boundaries) = if events.is_empty() {
-        (messages_to_items(&msgs), HashMap::new())
-    } else {
-        let first_seq = events.iter().find_map(|event| match event {
-            AgentEvent::MessageBoundary { seq, .. } => Some(*seq),
-            _ => None,
-        });
-        let prefix_len = first_seq.map_or(msgs.len(), |first_seq| {
-            page.messages
-                .iter()
-                .take_while(|(seq, _)| *seq < first_seq)
-                .count()
-        });
-        let mut prefix = messages_to_items(&msgs[..prefix_len]);
-        let prefix_items = prefix.len();
-        let (event_items, event_boundaries) = events_to_items(&events);
-        prefix.extend(event_items);
-        (
-            prefix,
-            event_boundaries
-                .into_iter()
-                .map(|(seq, offset)| (seq, prefix_items + offset))
-                .collect(),
-        )
-    };
-    let mut resources_by_seq = HashMap::<i64, Vec<resource_refs::UiMessageResource>>::new();
-    for resource in &page.resources {
-        resources_by_seq
-            .entry(resource.message_seq)
-            .or_default()
-            .push(resource.into());
-    }
-    for (message_seq, resources) in resources_by_seq {
-        let end = boundaries.get(&message_seq).copied().unwrap_or_else(|| {
-            let message_count = page
-                .messages
-                .iter()
-                .take_while(|(seq, _)| *seq <= message_seq)
-                .count();
-            messages_to_items(&msgs[..message_count]).len()
-        });
-        let end = end.min(items.len());
-        if let Some(item) = items[..end]
-            .iter_mut()
-            .rev()
-            .find(|item| item.role == "assistant")
-        {
-            item.resources = resources;
-        }
-    }
-    let mut inserted = 0usize;
-    for (message_seq, report_json) in &page.reviews {
-        let report: review::ReviewReport = serde_json::from_str(&report_json)
-            .map_err(|e| format!("invalid persisted review: {e}"))?;
-        let at = boundaries.get(message_seq).copied().unwrap_or_else(|| {
-            let message_count = page
-                .messages
-                .iter()
-                .take_while(|(seq, _)| seq <= message_seq)
-                .count();
-            messages_to_items(&msgs[..message_count]).len()
-        }) + inserted;
-        items.insert(
-            at,
-            UiItem {
-                role: "review".into(),
-                text: serde_json::to_string(&report).map_err(|e| format!("{e}"))?,
-                tool_name: None,
-                ok: None,
-                duration_ms: None,
-                input: None,
-                model_name: None,
-                call_id: None,
-                kind: None,
-                status: None,
-                locations: None,
-                resources: Vec::new(),
-            },
-        );
-        inserted += 1;
-    }
-    Ok(items)
-}
-
-#[tauri::command]
-async fn load_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-    before_seq: Option<i64>,
-) -> Result<SessionTranscriptPage, String> {
-    let page = state
-        .store
-        .load_session_transcript_page(&id, before_seq, SESSION_TRANSCRIPT_PAGE_TURNS)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    let presentations = if before_seq.is_none() {
-        state
-            .store
-            .load_latest_session_ui_event(&id, "ToolPresentation")
-            .await
-            .map_err(|e| format!("{e}"))?
-            .map(|json| serde_json::from_str::<AgentEvent>(&json))
-            .transpose()
-            .map_err(|e| format!("invalid persisted tool presentation: {e}"))?
-            .and_then(|event| match event {
-                AgentEvent::ToolPresentation {
-                    presentation_id,
-                    presentation_kind,
-                    payload,
-                    ..
-                } => Some(SessionPresentation {
-                    presentation_id,
-                    presentation_kind,
-                    payload,
-                }),
-                _ => None,
-            })
-            .into_iter()
-            .collect()
-    } else {
-        Vec::new()
-    };
-    let outline = if before_seq.is_none() {
-        state
-            .store
-            .load_session_user_messages(&id)
-            .await
-            .map_err(|e| format!("{e}"))?
-            .into_iter()
-            .enumerate()
-            .map(|(user_index, (seq, text))| SessionOutlineItem {
-                user_index,
-                seq,
-                text,
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-    if before_seq.is_none() {
-        state.set_active_frame(window.label(), Some(id.clone()));
-        let _ = state.store.mark_frame_seen(&id).await;
-        if let Some(rt) = state.sessions.lock().await.get(&id).cloned() {
-            rt.set_last_seq(page.latest_seq);
-        }
-    }
-    let items = transcript_page_items(&page)?;
-    Ok(SessionTranscriptPage {
-        items,
-        next_before_seq: page.next_before_seq,
-        user_offset: page.user_offset,
-        outline,
-        presentations,
-    })
-}
-
-/// Mark which session this window is viewing without loading it. The UI calls
-/// this instead of `load_session` when switching to a *running* session (it
-/// renders the cached streaming transcript), so uploads still attach to the
-/// viewed session (#194) — `load_session` would clobber the runtime's
-/// `last_seq` with the DB snapshot mid-stream.
-#[tauri::command]
-async fn set_viewed_session(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    id: String,
-) -> Result<(), String> {
-    state.set_active_frame(window.label(), Some(id.clone()));
-    let _ = state.store.mark_frame_seen(&id).await;
-    Ok(())
-}
-
-#[tauri::command]
-async fn list_artifacts(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    session_id: Option<String>,
-) -> Result<Vec<ArtifactInfo>, String> {
-    let frame_id = match session_id.as_deref().filter(|s| !s.is_empty()) {
-        Some(id) => Some(id.to_string()),
-        None => state.active_frame(window.label()),
-    };
-    let Some(fid) = frame_id else {
-        return Ok(vec![]);
-    };
-    let rows = state
-        .store
-        .list_artifacts(&fid)
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(rows
-        .into_iter()
-        .map(|(id, name, ct, path, ts)| ArtifactInfo {
-            id,
-            name: name.clone(),
-            kind: ct,
-            path,
-            ts,
-            project_id: None,
-            project_name: None,
-            session_id: None,
-            session_title: None,
-            size_bytes: None,
-            origin: None,
-        })
-        .collect())
-}
-
-#[tauri::command]
-async fn search_artifacts(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    query: Option<String>,
-    limit: Option<i64>,
-    project_id: Option<String>,
-    all_projects: Option<bool>,
-) -> Result<Vec<ArtifactInfo>, String> {
-    let ap = state.active(window.label());
-    let project_id = if all_projects.unwrap_or(false) {
-        None
-    } else {
-        project_id.as_deref().or(Some(ap.id.as_str()))
-    };
-    let rows = state
-        .store
-        .search_artifacts(
-            project_id,
-            query.as_deref().unwrap_or(""),
-            limit.unwrap_or(12),
-            None,
-        )
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(rows
-        .into_iter()
-        .map(|a| ArtifactInfo {
-            id: a.id,
-            name: a.name,
-            kind: a.kind,
-            path: a.path,
-            ts: a.ts,
-            project_id: Some(a.project_id),
-            project_name: Some(a.project_name),
-            session_id: Some(a.session_id),
-            session_title: Some(a.session_title),
-            size_bytes: a.size_bytes,
-            origin: Some(a.origin),
-        })
-        .collect())
-}
-
-#[tauri::command]
-async fn search_sessions(
-    state: State<'_, AppState>,
-    query: Option<String>,
-    limit: Option<i64>,
-    project_id: Option<String>,
-) -> Result<Vec<SessionSearchInfo>, String> {
-    let running = state.running_turns.lock().await.clone();
-    let awaiting = state.awaiting_confirm.lock().unwrap().clone();
-    let rows = state
-        .store
-        .search_sessions(
-            project_id.as_deref(),
-            query.as_deref().unwrap_or(""),
-            limit.unwrap_or(12),
-            None,
-        )
-        .await
-        .map_err(|e| format!("{e}"))?;
-    Ok(rows
-        .into_iter()
-        .map(|s| SessionSearchInfo {
-            status: session_runtime_status(
-                &s.id,
-                s.last_role.as_deref(),
-                s.unseen,
-                &running,
-                &awaiting,
-            )
-            .into(),
-            id: s.id,
-            project_id: s.project_id,
-            project_name: s.project_name,
-            title: s.title,
-            ts: s.created_at,
-            activity_at: s.activity_at,
-        })
-        .collect())
-}
-
-/// Given candidate artifact file paths (as they appear in chat), return the
-/// subset that can't be previewed: resolved against the project root and
-/// missing on disk, or outside the root. The UI drops these so a stale
-/// intermediate file doesn't linger as an artifact that 404s on click (#41).
-#[tauri::command]
-fn missing_files(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    paths: Vec<String>,
-) -> Result<Vec<String>, String> {
-    let ap = state.active(window.label());
-    Ok(paths
-        .into_iter()
-        .filter(|p| {
-            wisp_tools::safety::validate_file_path(&ap.root, p)
-                .map(|real| !real.exists())
-                .unwrap_or(true)
-        })
-        .collect())
-}
-
-#[tauri::command]
-async fn read_artifact(state: State<'_, AppState>, id: String) -> Result<FileContent, String> {
-    let row = state
-        .store
-        .get_artifact_detail(&id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| format!("artifact '{id}' not found"))?;
-    let root = PathBuf::from(row.project_root);
-    tokio::task::spawn_blocking(move || read_file_at(&root, row.path, None))
-        .await
-        .map_err(|e| format!("{e}"))?
-}
-
-#[tauri::command]
-async fn read_artifact_bytes(
-    state: State<'_, AppState>,
-    id: String,
-    max_bytes: Option<u64>,
-) -> Result<Response, String> {
-    let row = state
-        .store
-        .get_artifact_detail(&id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| format!("artifact '{id}' not found"))?;
-    let root = PathBuf::from(row.project_root);
-    let bytes =
-        tokio::task::spawn_blocking(move || read_file_bytes_at(&root, &row.path, max_bytes))
-            .await
-            .map_err(|e| format!("{e}"))??;
-    Ok(Response::new(bytes))
-}
-
-/// Read the immutable artifact version captured by a message resource binding.
-/// Resource previews must never follow the artifact's mutable latest-version pointer.
-#[tauri::command]
-async fn read_artifact_version(
-    state: State<'_, AppState>,
-    version_id: String,
-) -> Result<FileContent, String> {
-    let version = state
-        .store
-        .get_artifact_version(&version_id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| format!("artifact version '{version_id}' not found"))?;
-    let artifact = state
-        .store
-        .get_artifact_detail(&version.artifact_id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| format!("artifact '{}' not found", version.artifact_id))?;
-    let root = PathBuf::from(artifact.project_root);
-    tokio::task::spawn_blocking(move || read_file_at(&root, version.storage_path, None))
-        .await
-        .map_err(|e| format!("{e}"))?
-}
-
-#[tauri::command]
-async fn read_artifact_version_bytes(
-    state: State<'_, AppState>,
-    version_id: String,
-    max_bytes: Option<u64>,
-) -> Result<Response, String> {
-    let version = state
-        .store
-        .get_artifact_version(&version_id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| format!("artifact version '{version_id}' not found"))?;
-    let artifact = state
-        .store
-        .get_artifact_detail(&version.artifact_id)
-        .await
-        .map_err(|e| format!("{e}"))?
-        .ok_or_else(|| format!("artifact '{}' not found", version.artifact_id))?;
-    let root = PathBuf::from(artifact.project_root);
-    let bytes = tokio::task::spawn_blocking(move || {
-        read_file_bytes_at(&root, &version.storage_path, max_bytes)
-    })
-    .await
-    .map_err(|e| format!("{e}"))??;
-    Ok(Response::new(bytes))
-}
-
 fn mcp_lib_dir(_root: &std::path::Path) -> Option<PathBuf> {
     wisp_paths::bio_tools_dir().map(|d| d.join("lib"))
 }
@@ -7296,325 +5577,6 @@ async fn build_project_info(state: &AppState, label: &str) -> ProjectInfo {
     }
 }
 
-#[tauri::command]
-async fn get_project_info(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<ProjectInfo, String> {
-    Ok(build_project_info(&state, window.label()).await)
-}
-
-#[tauri::command]
-async fn get_capabilities(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<Capabilities, String> {
-    let ap = state.active(window.label());
-    let project = build_project_info(&state, window.label()).await;
-    let tags = load_skill_tags(&state.store).await;
-    let enabled = effective_enabled_skill_names(&state.store, &ap).await;
-    let skills = skill_infos(&ap.skills, &tags, enabled.as_ref());
-    Ok(Capabilities {
-        skills,
-        mcp_servers: list_mcp_servers(&ap.root),
-        memory_files: list_memory_files(&ap.memory),
-        project,
-    })
-}
-
-#[derive(Serialize, Clone)]
-struct MemoryView {
-    enabled: bool,
-    today_file: String,
-    files: Vec<MemoryFile>,
-}
-
-#[tauri::command]
-fn list_memory(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<Vec<MemoryFile>, String> {
-    let ap = state.active(window.label());
-    Ok(list_memory_files(&ap.memory))
-}
-
-#[tauri::command]
-async fn get_memory_view(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<MemoryView, String> {
-    let ap = state.active(window.label());
-    Ok(MemoryView {
-        enabled: load_memory_enabled(&state.store).await,
-        today_file: chrono::Local::now().format("%Y-%m-%d.md").to_string(),
-        files: list_memory_files(&ap.memory),
-    })
-}
-
-#[tauri::command]
-async fn set_memory_enabled(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    enabled: bool,
-) -> Result<MemoryView, String> {
-    save_memory_enabled(&state.store, enabled).await?;
-    clear_idle_agents(&state).await;
-    let ap = state.active(window.label());
-    Ok(MemoryView {
-        enabled,
-        today_file: chrono::Local::now().format("%Y-%m-%d.md").to_string(),
-        files: list_memory_files(&ap.memory),
-    })
-}
-
-#[tauri::command]
-async fn get_auto_review_enabled(state: State<'_, AppState>) -> Result<bool, String> {
-    Ok(load_auto_review_enabled(&state.store).await)
-}
-
-#[tauri::command]
-async fn set_auto_review_enabled(
-    state: State<'_, AppState>,
-    enabled: bool,
-) -> Result<bool, String> {
-    save_auto_review_enabled(&state.store, enabled).await?;
-    Ok(enabled)
-}
-
-#[tauri::command]
-async fn get_update_check_enabled(state: State<'_, AppState>) -> Result<bool, String> {
-    Ok(load_update_check_enabled(&state.store).await)
-}
-
-#[tauri::command]
-async fn set_update_check_enabled(
-    state: State<'_, AppState>,
-    enabled: bool,
-) -> Result<bool, String> {
-    save_update_check_enabled(&state.store, enabled).await?;
-    Ok(enabled)
-}
-
-#[tauri::command]
-fn read_memory_file(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    name: String,
-) -> Result<String, String> {
-    let ap = state.active(window.label());
-    let path = memory_file_path(&ap.memory, &name)?;
-    if !path.is_file() {
-        return Ok(String::new());
-    }
-    std::fs::read_to_string(&path).map_err(|e| format!("{e}"))
-}
-
-#[tauri::command]
-fn write_memory_file(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    name: String,
-    content: String,
-) -> Result<Vec<MemoryFile>, String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let path = memory_file_path(&ap.memory, &name)?;
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent).map_err(|e| format!("{e}"))?;
-    }
-    std::fs::write(&path, content).map_err(|e| format!("{e}"))?;
-    Ok(list_memory_files(&ap.memory))
-}
-
-#[tauri::command]
-fn delete_memory_file(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    name: String,
-) -> Result<Vec<MemoryFile>, String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let path = memory_file_path(&ap.memory, &name)?;
-    if path.is_file() {
-        std::fs::remove_file(&path).map_err(|e| format!("{e}"))?;
-    }
-    Ok(list_memory_files(&ap.memory))
-}
-
-#[tauri::command]
-fn clear_memory(
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-) -> Result<Vec<MemoryFile>, String> {
-    let ap = state.active(window.label());
-    let _project_activity = state.begin_project_activity(&ap.id)?;
-    let Ok(rd) = std::fs::read_dir(ap.memory.dir()) else {
-        return Ok(vec![]);
-    };
-    for entry in rd.flatten() {
-        let path = entry.path();
-        if path.extension().and_then(|e| e.to_str()) == Some("md") {
-            let _ = std::fs::remove_file(path);
-        }
-    }
-    Ok(list_memory_files(&ap.memory))
-}
-
-#[tauri::command]
-async fn get_onboarding_state(state: State<'_, AppState>) -> Result<OnboardingState, String> {
-    let (_, _, _, api_key) = load_settings(&state.store).await;
-    let done = state
-        .store
-        .get_setting("onboarding_done")
-        .await
-        .ok()
-        .flatten()
-        .is_some();
-    Ok(OnboardingState {
-        show: !done,
-        has_api_key: !api_key.is_empty(),
-    })
-}
-
-fn initial_bootstrap(workspace: &std::path::Path, skills: usize) -> BootstrapStatus {
-    let mut status = BootstrapStatus {
-        skills_loaded: skills,
-        python_ok: false,
-        python_initializing: true,
-        mcp_catalog: list_mcp_servers(workspace).len(),
-        uv_ok: wisp_runtime::PythonEnv::find_uv().is_some(),
-        node_ok: wisp_runtime::PythonEnv::find_node().is_some(),
-        npm_ok: wisp_runtime::PythonEnv::find_npm().is_some(),
-        sci_ok: wisp_runtime::PythonEnv::find_sci().is_some(),
-        pixi_ok: wisp_runtime::PythonEnv::find_pixi().is_some(),
-        app_version: env!("CARGO_PKG_VERSION").into(),
-        workspace: workspace.to_string_lossy().into_owned(),
-        errors: vec![],
-    };
-    if status.skills_loaded == 0 {
-        status
-            .errors
-            .push("No bundled skills found in install resources.".into());
-    }
-    if !status.uv_ok {
-        status
-            .errors
-            .push("uv not found on PATH; install uv or set UV_PATH.".into());
-    }
-    if !status.node_ok {
-        status
-            .errors
-            .push("Node.js not found on PATH; bear-* literature skills need Node >= 20.".into());
-    } else if !status.npm_ok {
-        status.errors.push(
-            "npm not found on PATH; install Node.js (includes npm) for scimaster-cli.".into(),
-        );
-    } else if !status.sci_ok {
-        status.errors.push(
-            "scimaster-cli (`sci`) not found; run `npm install -g scimaster-cli` then `sci init`."
-                .into(),
-        );
-    }
-    if !status.pixi_ok {
-        status.errors.push(
-            "pixi not found on PATH; optional for local bioinformatics multi-env workflows.".into(),
-        );
-    }
-    if wisp_paths::bio_tools_dir().is_none() {
-        status
-            .errors
-            .push("Bundled bio-tools MCP catalog not found.".into());
-    }
-    status
-}
-
-fn finish_python_bootstrap(status: &mut BootstrapStatus, result: Result<(), String>) {
-    status.python_initializing = false;
-    match result {
-        Ok(()) => status.python_ok = true,
-        Err(error) => status.errors.push(format!("Python environment: {error}")),
-    }
-}
-
-fn start_python_bootstrap(app: &tauri::AppHandle) {
-    let handle = app.clone();
-    let app_data = app.state::<AppState>().app_data.clone();
-    tauri::async_runtime::spawn(async move {
-        // Environment creation invokes uv and may download/install large wheels.
-        // Keep all of it off Tauri's event-loop thread so the first window stays
-        // responsive while the one-time bootstrap runs.
-        let result = tokio::task::spawn_blocking(move || {
-            wisp_runtime::PythonEnv::ensure(&app_data)
-                .map(|_| ())
-                .map_err(|error| error.to_string())
-        })
-        .await
-        .unwrap_or_else(|error| Err(format!("bootstrap task failed: {error}")));
-
-        let status = {
-            let state = handle.state::<AppState>();
-            let mut status = state.bootstrap.lock().unwrap();
-            finish_python_bootstrap(&mut status, result);
-            status.clone()
-        };
-        let _ = handle.emit("bootstrap-status", status);
-    });
-}
-
-#[tauri::command]
-fn get_bootstrap_status(state: State<'_, AppState>) -> BootstrapStatus {
-    state.bootstrap.lock().unwrap().clone()
-}
-
-#[tauri::command]
-fn open_external_url(app: tauri::AppHandle, url: String) -> Result<(), String> {
-    use tauri_plugin_opener::OpenerExt;
-    app.opener()
-        .open_url(url, None::<&str>)
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-fn reveal_in_file_manager(
-    app: AppHandle,
-    state: State<'_, AppState>,
-    window: tauri::WebviewWindow,
-    path: String,
-) -> Result<(), String> {
-    use tauri_plugin_opener::OpenerExt;
-    let ap = state.active(window.label());
-    let real = wisp_tools::safety::validate_file_path(&ap.root, &path)?;
-    if !real.exists() {
-        return Err(format!("file not found: {path}"));
-    }
-    app.opener()
-        .reveal_item_in_dir(&real)
-        .map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-async fn check_for_updates() -> Result<UpdateCheck, String> {
-    const LATEST_RELEASE_API: &str =
-        "https://api.github.com/repos/xuzhougeng/wisp-science/releases/latest";
-
-    let release = reqwest::Client::builder()
-        .timeout(std::time::Duration::from_secs(8))
-        .build()
-        .map_err(|error| format!("Failed to create update client: {error}"))?
-        .get(LATEST_RELEASE_API)
-        .header(reqwest::header::USER_AGENT, "wisp-science-update-check")
-        .send()
-        .await
-        .map_err(|error| format!("Failed to check GitHub Releases: {error}"))?
-        .error_for_status()
-        .map_err(|error| format!("GitHub Releases returned an error: {error}"))?
-        .json::<GithubRelease>()
-        .await
-        .map_err(|error| format!("Invalid response from GitHub Releases: {error}"))?;
-
-    update_check_from_release(env!("CARGO_PKG_VERSION"), release)
-}
-
 #[derive(Deserialize)]
 struct GithubRelease {
     tag_name: String,
@@ -7654,15 +5616,6 @@ fn update_check_from_release(
         release_url: release.html_url,
         notes: release.body,
     })
-}
-
-#[tauri::command]
-async fn dismiss_onboarding(state: State<'_, AppState>) -> Result<(), String> {
-    state
-        .store
-        .set_setting("onboarding_done", "1")
-        .await
-        .map_err(|e| format!("{e}"))
 }
 
 /// Tell the webview whether we're in dev (keep native context menu / DevTools).
@@ -7842,7 +5795,7 @@ pub fn run() {
 
             let skills = Arc::new(SkillIndex::load(&skill_paths(&root)));
             let memory = Arc::new(MemoryManager::new(&root));
-            let bootstrap = StdMutex::new(initial_bootstrap(&root, skills.all().len()));
+            let bootstrap = StdMutex::new(app_commands::initial_bootstrap(&root, skills.all().len()));
             let approvals = Arc::new(StdRwLock::new(tauri::async_runtime::block_on(
                 build_approval_policy(&store),
             )));
@@ -7911,7 +5864,7 @@ pub fn run() {
                     channels::autostart(handle).await;
                 });
             }
-            start_python_bootstrap(app.handle());
+            app_commands::start_python_bootstrap(app.handle());
             set_dev_flag(app.handle());
             #[cfg(target_os = "windows")]
             {
@@ -7948,13 +5901,13 @@ pub fn run() {
             {
                 let handle = app.handle().clone();
                 let ids = tauri::async_runtime::block_on(async {
-                    persisted_windows(&handle.state::<AppState>().store).await
+                    project_commands::persisted_windows(&handle.state::<AppState>().store).await
                 });
                 for id in ids {
                     let handle = app.handle().clone();
                     tauri::async_runtime::block_on(async {
                         let st = handle.state::<AppState>();
-                        let _ = spawn_project_window(&handle, st.inner(), &id, None).await;
+                        let _ = project_commands::spawn_project_window(&handle, st.inner(), &id, None).await;
                     });
                 }
             }
@@ -8025,36 +5978,36 @@ pub fn run() {
             terminal_sessions::write_terminal,
             terminal_sessions::resize_terminal,
             terminal_sessions::terminate_terminal,
-            new_session,
-            branch_session,
-            list_sessions,
-            list_sessions_page,
-            list_execution_contexts,
-            list_runtimes,
-            inspect_runtime,
-            execute_runtime,
-            start_runtime,
-            stop_runtime,
-            restart_runtime,
-            list_runs,
-            get_run,
-            cancel_run,
-            get_research_graph,
-            delete_session,
-            rename_session,
-            set_session_pinned,
-            transfer_session_to_project,
-            list_folders,
-            create_folder,
-            rename_folder,
-            delete_folder,
-            move_session,
-            list_recent_sessions,
-            list_projects,
-            pick_directory,
-            download_file,
-            export_text_file,
-            import_json_file,
+            session_commands::new_session,
+            session_commands::branch_session,
+            session_commands::list_sessions,
+            session_commands::list_sessions_page,
+            runtime_commands::list_execution_contexts,
+            runtime_commands::list_runtimes,
+            runtime_commands::inspect_runtime,
+            runtime_commands::execute_runtime,
+            runtime_commands::start_runtime,
+            runtime_commands::stop_runtime,
+            runtime_commands::restart_runtime,
+            runtime_commands::list_runs,
+            runtime_commands::get_run,
+            runtime_commands::cancel_run,
+            project_commands::get_research_graph,
+            session_commands::delete_session,
+            session_commands::rename_session,
+            session_commands::set_session_pinned,
+            session_commands::transfer_session_to_project,
+            session_commands::list_folders,
+            session_commands::create_folder,
+            session_commands::rename_folder,
+            session_commands::delete_folder,
+            session_commands::move_session,
+            session_commands::list_recent_sessions,
+            project_commands::list_projects,
+            app_commands::pick_directory,
+            app_commands::download_file,
+            app_commands::export_text_file,
+            app_commands::import_json_file,
             export_session,
             debug_request::export_debug_request,
             project_transfer::export_project,
@@ -8070,14 +6023,14 @@ pub fn run() {
             project_sync::project_sync_code,
             project_sync::join_synced_project,
             project_sync::get_project_sync_status,
-            create_project,
-            open_project,
-            open_project_window,
-            delete_project,
-            get_project_settings,
-            update_project,
-            load_session,
-            rewind_session,
+            project_commands::create_project,
+            project_commands::open_project,
+            project_commands::open_project_window,
+            project_commands::delete_project,
+            project_commands::get_project_settings,
+            project_commands::update_project,
+            session_commands::load_session,
+            session_commands::rewind_session,
             skill_commands::list_skills,
             skill_commands::set_skill_tags,
             skill_commands::set_skills_enabled,
@@ -8129,15 +6082,15 @@ pub fn run() {
             read_file_bytes,
             write_file,
             append_review_note,
-            list_artifacts,
-            search_artifacts,
-            search_sessions,
-            read_artifact,
-            read_artifact_bytes,
-            read_artifact_version,
-            read_artifact_version_bytes,
-            missing_files,
-            set_viewed_session,
+            artifact_commands::list_artifacts,
+            artifact_commands::search_artifacts,
+            session_commands::search_sessions,
+            artifact_commands::read_artifact,
+            artifact_commands::read_artifact_bytes,
+            artifact_commands::read_artifact_version,
+            artifact_commands::read_artifact_version_bytes,
+            artifact_commands::missing_files,
+            session_commands::set_viewed_session,
             upload_file,
             register_artifact,
             save_workspace_file_by_kind,
@@ -8150,26 +6103,26 @@ pub fn run() {
             library_commands::update_library_code,
             library_commands::list_library_item_versions,
             library_commands::delete_library_item,
-            get_project_info,
-            get_capabilities,
-            list_memory,
-            get_memory_view,
-            set_memory_enabled,
-            get_auto_review_enabled,
-            set_auto_review_enabled,
-            get_update_check_enabled,
-            set_update_check_enabled,
-            notify_user,
-            read_memory_file,
-            write_memory_file,
-            delete_memory_file,
-            clear_memory,
-            get_onboarding_state,
-            dismiss_onboarding,
-            get_bootstrap_status,
-            check_for_updates,
-            open_external_url,
-            reveal_in_file_manager,
+            project_commands::get_project_info,
+            app_commands::get_capabilities,
+            memory_commands::list_memory,
+            memory_commands::get_memory_view,
+            memory_commands::set_memory_enabled,
+            settings_commands::get_auto_review_enabled,
+            settings_commands::set_auto_review_enabled,
+            settings_commands::get_update_check_enabled,
+            settings_commands::set_update_check_enabled,
+            app_commands::notify_user,
+            memory_commands::read_memory_file,
+            memory_commands::write_memory_file,
+            memory_commands::delete_memory_file,
+            memory_commands::clear_memory,
+            app_commands::get_onboarding_state,
+            app_commands::dismiss_onboarding,
+            app_commands::get_bootstrap_status,
+            app_commands::check_for_updates,
+            app_commands::open_external_url,
+            app_commands::reveal_in_file_manager,
             connector_commands::list_mcp_connections,
             connector_commands::add_mcp_connection,
             connector_commands::authorize_http_connection,
